hm4, terminal, move exceptions from server to terminal and change float to int

+++ HomeWork4/src/ex01/Terminal.java
@@ -2,6 +2,6 @@ package ex01;
 
 public interface Terminal {
     void checkBalance();
-    void put(float amount);
-    void withdraw(float amount);
+    void put(int amount);
+    void withdraw(int amount);
 }
